@@ -8,9 +8,9 @@
     "build:mobile": "NODE_ENV=mobile:production npm run commons:build",
     "clean:browser": "rm -rf build/*",
     "clean:mobile": "rm -rf mobile/www/*",
-    "commons:build": "webpack --config webpack.config.js --bail",
-    "commons:watch": "webpack --config webpack.config.js --watch --display-modules --display-chunks",
-    "commons:watch:standalone": "webpack-dev-server --config webpack.config.js --port 8282 --display-modules --display-chunks --inline --hot",
+    "commons:build": "cozy-scripts build",
+    "commons:watch": "cozy-scripts watch",
+    "commons:watch:standalone": "cozy-scripts standalone",
     "deploy": "git-directory-deploy --username Cozy --email contact@cozycloud.cc --directory build/ --repo=https://$GITHUB_TOKEN@github.com/cozy/cozy-collect.git",
     "tx": "tx pull --all || true",
     "precommons:build": "npm-run-all lint tx",
@@ -53,6 +53,7 @@
   },
   "homepage": "https://github.com/cozy/cozy-collect#readme",
   "dependencies": {
+    "cozy-scripts": "^0.4.5",
     "date-fns": "^1.28.5",
     "eslint": "^4.9.0",
     "eslint-config-prettier": "^2.6.0",
@@ -68,7 +69,6 @@
     "husky": "^0.14.3",
     "intro.js": "2.8.0-alpha.1",
     "lint-staged": "^4.3.0",
-    "node-polyglot": "2.0.0",
     "piwik-react-router": "^0.8.2",
     "preact": "^8.1.0",
     "preact-compat": "^3.16.0",
@@ -82,10 +82,9 @@
     "timeago-react": "^1.0.7"
   },
   "devDependencies": {
-    "autoprefixer": "6.5.1",
     "babel-core": "6.18.0",
     "babel-eslint": "8.0.1",
-    "babel-loader": "6.2.5",
+    "babel-loader": "^7.1.2",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-react-jsx": "^6.24.1",
@@ -95,40 +94,23 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "babel-runtime": "6.18.0",
-    "browser-sync": "2.17.5",
-    "browser-sync-webpack-plugin": "1.1.3",
-    "copy-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "^4.3.1",
     "cozy-bar": "^4.8.5",
     "cozy-client-js": "^0.4.5",
     "cozy-ui": "7.0.1",
-    "css-loader": "0.25.0",
-    "css-mqpacker": "5.0.1",
-    "csswring": "^5.1.1",
     "enzyme": "^2.9.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
+    "eslint-loader": "^1.9.0",
     "git-directory-deploy": "^1.5.1",
-    "html-webpack-plugin": "^2.28.0",
     "identity-obj-proxy": "^3.0.0",
-    "imports-loader": "^0.7.1",
     "jest": "^18.0.0",
-    "json-loader": "0.5.4",
     "npm-run-all": "^4.0.2",
-    "postcss": "5.2.5",
-    "postcss-assets-webpack-plugin": "^1.0.0",
-    "postcss-loader": "^1.3.3",
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
     "react-test-renderer": "^15.4.1",
-    "script-ext-html-webpack-plugin": "^1.7.1",
-    "style-loader": "0.13.1",
     "stylint": "^1.5.9",
     "stylus": "0.54.5",
-    "stylus-loader": "2.3.1",
-    "svg-sprite-loader": "^0.3.0",
     "url-search-params-polyfill": "^1.2.0",
-    "webpack": "^1.14.0",
-    "webpack-merge": "^4.1.0"
+    "webpack": "^3.10.0"
   },
   "jest": {
     "moduleFileExtensions": [
