@@ -1,3 +1,3 @@
 'use strict'
 
-module.exports = 'test-file-stub'
+export default 'test-file-stub'
