@@ -1,7 +1,8 @@
 'use strict'
 
 const path = require('path')
-const { ProvidePlugin } = require('webpack')
+const { ProvidePlugin, DefinePlugin } = require('webpack')
+const CopyPlugin = require('copy-webpack-plugin')
 
 const SRC_DIR = path.resolve(__dirname, '../src')
 
@@ -14,8 +15,13 @@ module.exports = {
   },
   plugins: [
     new ProvidePlugin({
-      'initKonnectors': 'expose?initKonnectors!config/konnectors.json',
-      'initFolders': 'expose?intFolders!config/folders.json'
-    })
+      initKonnectors: 'expose-loader?initKonnectors!config/konnectors.json'
+    }),
+    new DefinePlugin({
+      __PIWIK_SITEID__: 8,
+      __PIWIK_DIMENSION_ID_APP__: 1,
+      __PIWIK_TRACKER_URL__: JSON.stringify('https://piwik.cozycloud.cc')
+    }),
+    new CopyPlugin([{ from: 'scripts/publish_registry.sh' }])
   ]
 }
