@@ -1,4 +1,3 @@
-/* global __DEVTOOLS__ */
 import { compose, createStore, applyMiddleware } from 'redux'
 import { cozyMiddleware } from 'redux-cozy-client'
 import { createLogger } from 'redux-logger'
@@ -7,16 +6,9 @@ import thunkMiddleware from 'redux-thunk'
 import CollectStore from '../lib/CollectStore'
 import getReducers from '../reducers'
 
-const configureStore = (
-  client,
-  initKonnectors,
-  initFolders,
-  context,
-  options = {}
-) => {
+const configureStore = (client, initKonnectors, context, options = {}) => {
   // Enable Redux dev tools
-  const composeEnhancers =
-    (__DEVTOOLS__ && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose
+  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE || compose
 
   const reduxStore = createStore(
     getReducers(),
@@ -30,7 +22,7 @@ const configureStore = (
   )
 
   return Object.assign(
-    new CollectStore(initKonnectors, initFolders, context, options),
+    new CollectStore(initKonnectors, context, options),
     reduxStore
   )
 }
